fix(ci): python-coverage-shards — split oversized auth_mounts_claude.py under the 1500-line maintainability limit
The #402/#404/#405 auth-overlay hardening grew
src/awf/node/auth_mounts_claude.py to 1628 lines, tripping the first-party
line-limit guardrail (test_first_party_code_files_stay_under_line_limit,
MAX_FIRST_PARTY_FILE_LINES = 1500) on coverage shard 8.

Decompose into two dependency-free leaf modules, preserving the strict import
DAG (no cycles) and the auth_mounts facade re-export surface:

- auth_mounts_caps.py: Linux CapEff probes (_capeff_has_bit,
  _has_cap_sys_admin, _has_cap_mknod) + their constants.
- auth_mounts_claude_reconcile.py: the fallback edit/deletion reconcile pass
  (_legacy_is_unedited_host_copy, _forward_fallback_deletions_as_whiteouts,
  _reconcile_fallback_edits_into_upper) + its constants.

auth_mounts_claude.py re-imports the externally/internally needed names so
awf.node.auth_mounts.<name> stays the stable import surface. The reconcile
tests that monkeypatch _has_cap_mknod/_safe_stat are retargeted to the module
that now owns those bindings. File is now 1282 lines; node suite, mypy, ruff
all green.
